% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{add_pipe}
\alias{add_pipe}
\title{Add a pipe operator to the last element of a character vector}
\usage{
add_pipe(code_block)
}
\arguments{
\item{code_block}{A character vector.}
}
\value{
The character vector with a pipe operator added to the last element.
}
\description{
Add a pipe operator to the last element of a character vector
}
\keyword{internal}
