% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{are_to_rewrite}
\alias{are_to_rewrite}
\title{Are values to be rewritten?}
\usage{
are_to_rewrite(x, from)
}
\arguments{
\item{x}{A vector of values that will be checked against the \code{from} vector.}

\item{from}{A vector of values that \code{x} will be checked for matches against.}
}
\value{
A logical vector of the same length as \code{x}, where \code{TRUE} indicates
that the corresponding value in \code{x} matches a value in \code{from} and
should be rewritten, and \code{FALSE} otherwise. If \code{x} is empty, returns
an empty logical vector. This function is intended for internal use
and optimization in data transformation processes.
}
\description{
\code{are_to_rewrite} is a helper function designed to determine if any values
in a vector \code{x} match the specified \code{from} values, indicating they are
candidates for recoding or rewriting.
}
\examples{
sdtm.oak:::are_to_rewrite(x = 1:5, from = c(2, 4))

sdtm.oak:::are_to_rewrite(letters[1:3], from = c("a", "c"))

}
\keyword{internal}
