% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{assert_ct}
\alias{assert_ct}
\title{Assert a controlled terminology specification}
\usage{
assert_ct(ct)
}
\arguments{
\item{ct}{A data frame to be asserted as a controlled terminology data set.}
}
\value{
The function throws an error if \code{ct} is not a valid controlled
terminology data set; otherwise, \code{ct} is returned invisibly.
}
\description{
\code{\link[=assert_ct]{assert_ct()}} will check whether \code{ct} is a data frame and
if it contains the variables: codelist_code, collected_value, term_synonyms, and term_value.
}
\examples{
# If `ct` is a valid controlled terminology then it is returned invisibly.
ct_01 <- read_ct_example("ct-01-cm")
all.equal(ct_01, sdtm.oak:::assert_ct(ct_01))

# A minimal set of variables needs to be present in `ct` for it to pass the
# assertion; `sdtm.oak:::ct_vars()` defines their names.
(req_vars <- sdtm.oak:::ct_vars())

# Other (facultative) variables also present in the controlled terminology
# example.
(opt_vars <- setdiff(colnames(ct_01), req_vars))

# With only the mandatory variables, the assertion still passes.
sdtm.oak:::assert_ct(ct_01[req_vars])

# Not having the required variables results in an error.
try(sdtm.oak:::assert_ct(ct_01[opt_vars]))

}
\keyword{internal}
