% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{assert_ct_cltc}
\alias{assert_ct_cltc}
\title{Assert a code-list code}
\usage{
assert_ct_cltc(ct_spec, ct_cltc, optional = FALSE)
}
\arguments{
\item{ct_spec}{Either a data frame encoding a controlled terminology data set, or
\code{NULL}.}

\item{ct_cltc}{A string with a to-be asserted code-list code, or \code{NULL}.}

\item{optional}{A scalar logical, indicating whether \code{ct_cltc} can be \code{NULL} or
not.}
}
\value{
The function throws an error if \code{ct_cltc} is not a valid code-list code
given the controlled terminology data set; otherwise, \code{ct_cltc} is returned
invisibly.
}
\description{
\code{\link[=assert_ct_cltc]{assert_ct_cltc()}} asserts the validity of a code-list code in the context of
a controlled terminology specification.
}
\examples{
# Load a controlled terminology example.
(ct_spec <- read_ct_spec_example("ct-01-cm"))

# Should work fine.
sdtm.oak:::assert_ct_cltc(ct_spec = ct_spec, ct_cltc = "C71113")

# In certain cases, you might allow `ct_cltc` to be `NULL` as to indicate absence,
# in that case, set `optional` to `TRUE` to make `assert_ct_cltc()` more
# forgiving.
sdtm.oak:::assert_ct_cltc(ct_spec = ct_spec, ct_cltc = NULL, optional = TRUE)

# Otherwise it would err.
try(sdtm.oak:::assert_ct_cltc(ct_spec = ct_spec, ct_cltc = NULL, optional = FALSE))

}
\keyword{internal}
