% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear \code{{sdtm.oak}} cache of memoised functions}
\usage{
clear_cache()
}
\value{
Returns a logical value, indicating whether the resetting of the
cache was successful (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Some of \code{{sdtm.oak}} functions have their results cached for runtime
efficiency. Use this function to reset the cache.

Memoised functions:
\itemize{
\item \code{\link[=ct_mappings]{ct_mappings()}}
}
}
\examples{
clear_cache()

}
