% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\docType{data}
\name{cm_template_prefix}
\alias{cm_template_prefix}
\title{The template suffix for the cm code}
\format{
An object of class \code{glue} (inherits from \code{character}) of length 1.
}
\usage{
cm_template_prefix
}
\description{
The template suffix for the cm code
}
\keyword{internal}
