% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{condition_by}
\alias{condition_by}
\title{Condition a data set based on specified conditions}
\usage{
condition_by(dat, ..., .na = NA, .env = rlang::env())
}
\arguments{
\item{dat}{A tibble.}

\item{...}{Conditions to filter the tibble.}

\item{.na}{Return value to be used when the conditions evalute to \code{NA}.}

\item{.env}{An optional environment to look for variables involved in logical
expression passed in \code{...}. A data frame or a list can also be passed that
will be coerced to an environment internally.}
}
\value{
A tibble with an additional class 'cnd_df' and a logical vector
attribute indicating matching rows.
}
\description{
This function tags records in a data set, indicating which rows match the
specified conditions.
}
