% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct_map}
\alias{ct_map}
\title{Recode according to controlled terminology}
\usage{
ct_map(x, ct = NULL, cl = NULL, from = ct_vars("from"), to = ct_vars("to"))
}
\arguments{
\item{x}{A character vector of terms to be recoded following a controlled
terminology.}

\item{ct}{A \link[tibble:tibble-package]{tibble} providing a controlled
terminology specification.}

\item{cl}{A character vector indicating a set of possible controlled
terminology code-lists codes to be used for recoding. By default (\code{NULL})
all code-lists available in \code{ct} are used.}

\item{from}{A character vector of column names indicating the variables
containing values to be matched against for terminology recoding.}

\item{to}{A single string indicating the column whose values are to be
recoded into.}
}
\value{
A character vector of terminology recoded values from \code{x}. If no
match is found in the controlled terminology spec provided in \code{ct}, then
\code{x} values are returned in uppercase. If \code{ct} is not provided \code{x} is
returned unchanged.
}
\description{
\code{\link[=ct_map]{ct_map()}} recodes a vector following a controlled terminology.
}
\examples{
# A few example terms.
terms <-
  c("/day",
    "Yes",
    "Unknown",
    "Prior",
    "Every 2 hours",
    "Percentage",
    "International Unit")

# Load a controlled terminology example
(ct <- read_ct_example("ct-01-cm"))

# Use all possible matching terms in the controlled terminology.
ct_map(x = terms, ct = ct)

# Note that if the controlled terminology mapping is restricted to a code-list
# code, e.g. C71113, then only `"/day"` gets mapped to `"QD"`; remaining terms
# won't match given the code-list code restriction, and will be mapped to an
# uppercase version of the original terms.
ct_map(x = terms, ct = ct, cl = "C71113")

}
