% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct_vars}
\alias{ct_vars}
\title{Controlled terminology variables}
\usage{
ct_vars(set = c("all", "cl", "from", "to"))
}
\arguments{
\item{set}{A scalar character (string), one of: \code{"all"} (default), \code{"cl"},
\code{"from"} or \code{"to"}.}
}
\description{
\code{\link[=ct_vars]{ct_vars()}} returns the mandatory variables to be present in a data set
representing a controlled terminology. By default, it returns all required
variables.

If only the subset of variables used for matching terms are needed, then
request this subset of variables by passing the argument value \code{"from"}. If
only the mapping-to variable is to be requested, then simply pass \code{"to"}. If
only the code-list code variable name is needed then pass \code{"cl"}.
}
\examples{
# These two calls are equivalent and return all required variables in a
# controlled terminology data set.
sdtm.oak:::ct_vars()
sdtm.oak:::ct_vars("all")

# "Codelist code" variable name.
sdtm.oak:::ct_vars("cl")

# "From" variables
sdtm.oak:::ct_vars("from")

# The "to" variable.
sdtm.oak:::ct_vars("to")

}
\keyword{internal}
