% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{ctl_new_rowid_pillar.cnd_df}
\alias{ctl_new_rowid_pillar.cnd_df}
\title{Conditioned tibble pillar print method}
\usage{
\method{ctl_new_rowid_pillar}{cnd_df}(controller, x, width, ...)
}
\arguments{
\item{controller}{The object of class \code{"tbl"} currently printed.}

\item{x}{A simple (one-dimensional) vector.}

\item{width}{The available width, can be a vector for multiple tiers.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A character vector to print the tibble which is a conditioned dataframe.
}
\description{
Conditioned tibble pillar print method
}
\seealso{
\code{\link[=tbl_sum.cnd_df]{tbl_sum.cnd_df()}}.
}
