% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_blfl.R
\name{dtc_datepart}
\alias{dtc_datepart}
\title{Extract date part from ISO8601 date/time variable}
\usage{
dtc_datepart(dtc, partial_as_na = TRUE)
}
\arguments{
\item{dtc}{Character vector containing ISO8601 date/times.}

\item{partial_as_na}{Logical \code{TRUE} or \code{FALSE} indicating whether
partial dates should be set to NA (default is \code{TRUE}).}
}
\value{
Character vector containing ISO8601 dates.
}
\description{
The date part is extracted from an ISO8601 date/time variable.
By default, partial or missing dates are set to NA.
}
\keyword{internal}
