% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_formats.R
\docType{data}
\name{dtc_formats}
\alias{dtc_formats}
\title{Date/time collection formats}
\format{
A \link[tibble:tibble-package]{tibble} of 20 formats
with three variables:
\describe{
\item{\code{fmt}}{Format string.}
\item{\code{type}}{Whether a date, time or date-time.}
\item{\code{description}}{Description of which date-time components are parsed.}
}
}
\usage{
dtc_formats
}
\description{
Date/time collection formats
}
\examples{
dtc_formats

}
\keyword{datasets}
