% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_blfl.R
\name{dtc_timepart}
\alias{dtc_timepart}
\title{Extract time part from ISO 8601 date/time variable}
\usage{
dtc_timepart(dtc, partial_as_na = TRUE, ignore_seconds = TRUE)
}
\arguments{
\item{dtc}{Character vector containing ISO 8601 date/times.}

\item{partial_as_na}{Logical \code{TRUE} or \code{FALSE} indicating whether
partial times should be set to NA (default is \code{TRUE}).}

\item{ignore_seconds}{Logical \code{TRUE} or \code{FALSE} indicating whether
seconds should be ignored (default is \code{TRUE}).}
}
\value{
Character vector containing ISO 8601 times.
}
\description{
The time part is extracted from an ISO 8601 date/time variable.
By default, partial or missing times are set to NA, and seconds are ignored
and not extracted.
}
\keyword{"",}
\keyword{"12",}
\keyword{"12:30")}
\keyword{"12:30",}
\keyword{"12:30:59")}
\keyword{"2021-12-25",}
\keyword{"2021-12-25T12",}
\keyword{"2021-12-25T12:30",}
\keyword{"2021-12-25T12:30:59")}
\keyword{"2021-12-25T12:30:59"),}
\keyword{#}
\keyword{##}
\keyword{)}
\keyword{=}
\keyword{Do}
\keyword{FALSE}
\keyword{FALSE,}
\keyword{NA}
\keyword{NA,}
\keyword{Partial}
\keyword{Prevent}
\keyword{and}
\keyword{being}
\keyword{by}
\keyword{c(NA,}
\keyword{default}
\keyword{from}
\keyword{ignore}
\keyword{ignore_seconds}
\keyword{ignored}
\keyword{internal}
\keyword{missing}
\keyword{not}
\keyword{or}
\keyword{partial}
\keyword{partial_as_na}
\keyword{prevent}
\keyword{sdtm.oak:::dtc_timepart(}
\keyword{seconds}
\keyword{seconds,}
\keyword{set}
\keyword{times}
\keyword{to}
\keyword{|-->}
