% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{dttm_fmt_to_regex}
\alias{dttm_fmt_to_regex}
\title{Convert a parsed date/time format to regex}
\usage{
dttm_fmt_to_regex(tbl_fmt_c, fmt_regex = fmt_rg(), anchored = TRUE)
}
\arguments{
\item{tbl_fmt_c}{A \link[tibble:tibble-package]{tibble} of parsed date/time
format components as returned by \code{\link[=parse_dttm_fmt]{parse_dttm_fmt()}}.}

\item{fmt_regex}{A named character vector of regexps, one for each date/time
component.}

\item{anchored}{Whether the final regex should be anchored, i.e. bounded by
\code{"^"} and \code{"$"} for a whole match.}
}
\value{
A string containing a regular expression for matching date/time
components according to a format.
}
\description{
\code{\link[=dttm_fmt_to_regex]{dttm_fmt_to_regex()}} takes a \link[tibble:tibble-package]{tibble} of parsed
date/time format components (as returned by \code{\link[=parse_dttm_fmt]{parse_dttm_fmt()}}), and a
mapping of date/time component formats to regexps and generates a single
regular expression with groups for matching each of the date/time components.
}
\examples{
tbl_fmt_c <- sdtm.oak:::parse_dttm_fmt("y")
sdtm.oak:::dttm_fmt_to_regex(tbl_fmt_c)
sdtm.oak:::dttm_fmt_to_regex(tbl_fmt_c, anchored = FALSE)

tbl_fmt_c <- sdtm.oak:::parse_dttm_fmt("m")
sdtm.oak:::dttm_fmt_to_regex(tbl_fmt_c)

tbl_fmt_c <- sdtm.oak:::parse_dttm_fmt("ymd")
sdtm.oak:::dttm_fmt_to_regex(tbl_fmt_c)

tbl_fmt_c <- sdtm.oak:::parse_dttm_fmt("ymd HH:MM:SS")
sdtm.oak:::dttm_fmt_to_regex(tbl_fmt_c)

}
\keyword{internal}
