% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\docType{data}
\name{expected_columns}
\alias{expected_columns}
\title{Expected columns in the specification file}
\format{
An object of class \code{character} of length 50.
}
\usage{
expected_columns
}
\description{
Expected columns in the specification file
}
\keyword{internal}
