% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_supp.R
\name{gen_supp}
\alias{gen_supp}
\title{Function to generate supplemental domains SUPP--}
\usage{
gen_supp(sdtm_dataset = NULL, idvar = NULL, spec = NULL, eval = NA_character_)
}
\arguments{
\item{sdtm_dataset}{SDTM output used to split supplemental domains.}

\item{idvar}{Variable name used to for variable.}

\item{spec}{Data frame of specifications..}

\item{eval}{Character string of the evaluator.}
}
\value{
Supplemental domains.
}
\description{
Function to generate supplemental domains SUPP--
}
\examples{
path <- here::here("inst")
dm <- readRDS(paste0(path, "/domain/dm.rds"))
spec <- read.csv(paste0(path, "/spec/spec.csv"))
suppdm <- gen_supp(dm, idvar = NULL, spec = spec, eval = "CLINICAL STUDY SPONSOR")
}
