% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{generate_code}
\alias{generate_code}
\title{Generate the code for the mapping SDTM specification}
\usage{
generate_code(spec, domain, out_dir = ".")
}
\arguments{
\item{spec}{The specification data frame.}

\item{domain}{The SDTM domain to generate the code for.}

\item{out_dir}{The directory to save the code file. Default is the current
directory.}
}
\value{
Side effect: the code is generated and saved to a file.
}
\description{
Generate the code for the mapping SDTM specification
}
\examples{

\dontrun{
spec <- read_spec("cm_sdtm_oak_spec_cdash.csv")
domain <- "cm"
generate_code(spec, domain)
}

}
