% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{generate_one_topic_code}
\alias{generate_one_topic_code}
\title{Generate the code for one topic}
\usage{
generate_one_topic_code(topic, domain, spec)
}
\arguments{
\item{topic}{The topic to generate the code for.}

\item{domain}{The SDTM domain.}

\item{spec}{The specification data frame.}
}
\value{
The code for the topic as a string.
}
\description{
Generate the code for one topic
}
\keyword{internal}
