% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{get_cnd_df_cnd}
\alias{get_cnd_df_cnd}
\title{Get the conditioning vector from a conditioned data frame}
\usage{
get_cnd_df_cnd(dat)
}
\arguments{
\item{dat}{A conditioned data frame (\code{cnd_df}).}
}
\value{
The conditioning vector (\code{cnd}) if \code{dat} is a conditioned data frame
(\code{cnd_df}), otherwise \code{NULL}, or \code{NULL} if \code{dat} is not a conditioned data
frame (\code{cnd_df}).
}
\description{
\code{\link[=get_cnd_df_cnd]{get_cnd_df_cnd()}} extracts the conditioning vector from a conditioned data
frame, i.e. from an object of class \code{cnd_df}.
}
\seealso{
\code{\link[=new_cnd_df]{new_cnd_df()}}, \code{\link[=is_cnd_df]{is_cnd_df()}}, \code{\link[=get_cnd_df_cnd_sum]{get_cnd_df_cnd_sum()}},
\code{\link[=rm_cnd_df]{rm_cnd_df()}}.
}
\keyword{internal}
