% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{get_cnd_df_cnd_sum}
\alias{get_cnd_df_cnd_sum}
\title{Get the summary of the conditioning vector from a conditioned data frame}
\usage{
get_cnd_df_cnd_sum(dat)
}
\arguments{
\item{dat}{A conditioned data frame (\code{cnd_df}).}
}
\value{
A vector of three elements providing the sum of \code{TRUE}, \code{FALSE}, and
\code{NA} values in the conditioning vector (\code{cnd}), or \code{NULL} if \code{dat} is not
a conditioned data frame (\code{cnd_df}).
}
\description{
\code{\link[=get_cnd_df_cnd_sum]{get_cnd_df_cnd_sum()}} extracts the tally of the conditioning vector from a
conditioned data frame.
}
\seealso{
\code{\link[=new_cnd_df]{new_cnd_df()}}, \code{\link[=is_cnd_df]{is_cnd_df()}}, \code{\link[=get_cnd_df_cnd]{get_cnd_df_cnd()}}, \code{\link[=rm_cnd_df]{rm_cnd_df()}}.
}
\keyword{internal}
