% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{get_domain_spec}
\alias{get_domain_spec}
\title{Get the specification for a domain and modify it}
\usage{
get_domain_spec(spec, domain)
}
\arguments{
\item{spec}{The specification data frame.}

\item{domain}{The SDTM domain to get the specification for.}
}
\value{
A tibble with the specification for the domain.
}
\description{
Get the specification for a domain and modify it
}
\keyword{internal}
