% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardcode.R
\name{harcode}
\alias{harcode}
\alias{hardcode_no_ct}
\alias{hardcode_ct}
\title{Derive an SDTM variable with a hardcoded value}
\usage{
hardcode_no_ct(
  raw_dat,
  raw_var,
  tgt_var,
  tgt_val,
  tgt_dat = NULL,
  id_vars = oak_id_vars()
)

hardcode_ct(
  raw_dat,
  raw_var,
  ct,
  cl,
  tgt_var,
  tgt_val,
  tgt_dat = NULL,
  id_vars = oak_id_vars()
)
}
\arguments{
\item{raw_dat}{The raw dataset.}

\item{raw_var}{The raw variable.}

\item{tgt_var}{The target SDTM variable.}

\item{tgt_val}{Hardcoded value.}

\item{tgt_dat}{Target dataset. By default the same as \code{raw_dataset}.}

\item{id_vars}{If \code{target_dataset} is different than \code{raw_dataset},
then this parameter defines keys to use in the join between \code{raw_dataset}
and \code{target_dataset}.}

\item{ct}{Study controlled terminology specification.}

\item{cl}{A codelist code indicating which
subset of the controlled terminology to apply in the derivation.}
}
\value{
The target dataset with the derived variable \code{target_sdtm_variable}.
}
\description{
\itemize{
\item \code{\link[=hardcode_no_ct]{hardcode_no_ct()}} maps a hardcoded value to a target SDTM variable that has
no terminology restrictions.
\item \code{\link[=hardcode_ct]{hardcode_ct()}} maps a hardcoded value to a target SDTM variable with
controlled terminology recoding.
}
}
\examples{
MD1 <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~MDRAW,
    1L, "MD1", 101L, "BABY ASPIRIN",
    2L, "MD1", 102L, "CORTISPORIN",
    3L, "MD1", 103L, NA_character_,
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS".
hardcode_no_ct(
  raw_dat = MD1,
  raw_var = "MDRAW",
  tgt_var = "CMCAT",
  tgt_val = "GENERAL CONCOMITANT MEDICATIONS"
)

CM_INTER <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~CMTRT, ~CMINDC,
    1L, "MD1", 101L, "BABY ASPIRIN", NA,
    2L, "MD1", 102L, "CORTISPORIN", "NAUSEA",
    3L, "MD1", 103L, "ASPIRIN", "ANEMIA",
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL", "NAUSEA",
    5L, "MD1", 105L, "PARACETAMOL", "PYREXIA"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS" with a prior join to
# `target_dataset`.
hardcode_no_ct(
  raw_dat = MD1,
  raw_var = "MDRAW",
  tgt_var = "CMCAT",
  tgt_val = "GENERAL CONCOMITANT MEDICATIONS",
  tgt_dat = CM_INTER,
  id_vars = c("oak_id", "raw_source", "patient_number")
)

}
