% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardcode_no_ct.R
\name{hardcode_no_ct}
\alias{hardcode_no_ct}
\title{Derive an SDTM variable with a hardcoded value}
\usage{
hardcode_no_ct(
  raw_dataset,
  raw_variable,
  target_sdtm_variable,
  target_hardcoded_value,
  target_dataset = raw_dataset,
  merge_to_topic_by = NULL
)
}
\arguments{
\item{raw_dataset}{The raw dataset.}

\item{raw_variable}{The raw variable.}

\item{target_sdtm_variable}{The target SDTM variable.}

\item{target_hardcoded_value}{Hardcoded value.}

\item{target_dataset}{Target dataset. By default the same as \code{raw_dataset}.}

\item{merge_to_topic_by}{If \code{target_dataset} is different than \code{raw_dataset},
then this parameter defines keys to use in the join between \code{raw_dataset}
and \code{target_dataset}.}
}
\description{
\code{\link[=hardcode_no_ct]{hardcode_no_ct()}} maps a hardcoded value to a target SDTM variable that has
no terminology restrictions.
}
\examples{
MD1 <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~MDRAW,
    1L, "MD1", 101L, "BABY ASPIRIN",
    2L, "MD1", 102L, "CORTISPORIN",
    3L, "MD1", 103L, NA_character_,
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS".
hardcode_no_ct(
  raw_dataset = MD1,
  raw_variable = "MDRAW",
  target_sdtm_variable = "CMCAT",
  target_hardcoded_value = "GENERAL CONCOMITANT MEDICATIONS"
)

CM_INTER <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~CMTRT, ~CMINDC,
    1L, "MD1", 101L, "BABY ASPIRIN", NA,
    2L, "MD1", 102L, "CORTISPORIN", "NAUSEA",
    3L, "MD1", 103L, "ASPIRIN", "ANEMIA",
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL", "NAUSEA",
    5L, "MD1", 105L, "PARACETAMOL", "PYREXIA"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS" with a prior join to
# `target_dataset`.

hardcode_no_ct(
  raw_dataset = MD1,
  raw_variable = "MDRAW",
  target_sdtm_variable = "CMCAT",
  target_hardcoded_value = "GENERAL CONCOMITANT MEDICATIONS",
  target_dataset = CM_INTER,
  merge_to_topic_by = c("oak_id", "raw_source", "patient_number")
)

}
