% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{index_for_rewrite}
\alias{index_for_rewrite}
\title{Determine Indices for Rewriting}
\usage{
index_for_rewrite(x, from)
}
\arguments{
\item{x}{A vector of values in which to search for matches.}

\item{from}{A vector of values to match against the elements in \code{x}.}
}
\value{
An integer vector of the same length as \code{x}, containing the indices
of the matched values from the \code{from} vector. If an element in \code{x} does not
match any value in \code{from}, the corresponding position in the output will be
\code{NA}. This index information is critical for subsequent rewrite operations.
}
\description{
\code{\link[=index_for_rewrite]{index_for_rewrite()}} identifies the positions of elements in \code{x} that match
any of the values specified in the \code{from} vector. This function is primarily
used to facilitate the rewriting of values by pinpointing which elements in
\code{x} correspond to the \code{from} values and thus need to be replaced or updated.
}
\examples{
sdtm.oak:::index_for_rewrite(x = 1:5, from = c(2, 4))

}
\keyword{internal}
