% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{is_character}
\alias{is_character}
\title{Check if a variable is character}
\usage{
is_character(var_in)
}
\arguments{
\item{var_in}{The variable to check.}
}
\value{
Logical indicating if the variable is character.
}
\description{
Check if a variable is character
}
