% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{is_cnd_df}
\alias{is_cnd_df}
\title{Check if a data frame is a conditioned data frame}
\usage{
is_cnd_df(dat)
}
\arguments{
\item{dat}{A data frame.}
}
\value{
\code{TRUE} if \code{dat} is a conditioned data frame (class \code{cnd_df}),
otherwise \code{FALSE}.
}
\description{
\code{\link[=is_cnd_df]{is_cnd_df()}} checks whether a data frame is a conditioned data frame, i.e.
of class \code{cnd_df}.
}
\seealso{
\code{\link[=new_cnd_df]{new_cnd_df()}}, \code{\link[=get_cnd_df_cnd]{get_cnd_df_cnd()}}, \code{\link[=get_cnd_df_cnd_sum]{get_cnd_df_cnd_sum()}},
\code{\link[=rm_cnd_df]{rm_cnd_df()}}.
}
\keyword{internal}
