% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_utils.R
\name{is_dtc_fmt}
\alias{is_dtc_fmt}
\title{Is it one of the supported formats?}
\usage{
is_dtc_fmt(fmt)
}
\arguments{
\item{fmt}{A character vector.}
}
\value{
A \link[base:logical]{logical} vector of the same size of \code{fmt}.
}
\description{
\code{\link[=is_dtc_fmt]{is_dtc_fmt()}} takes a character vector of date/time formats and returns
a logical indicating which are supported.
}
\details{
This function works by checking if the values of \code{x} are one of the formats
listed in column \code{fmt} of \link{dtc_formats}.
}
\examples{
sdtm.oak:::is_dtc_fmt(c("ymd", "y m d", "dmy", "HM", "H:M:S", "y-m-d H:M:S"))

sdtm.oak:::is_dtc_fmt("y years m months d days")

}
\keyword{internal}
