% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_blfl.R
\name{is_iso8601}
\alias{is_iso8601}
\title{This function is used to check if a --DTC variable is in ISO8601 format}
\usage{
is_iso8601(dtc_var)
}
\arguments{
\item{dtc_var}{A vector of the date and time values}
}
\value{
A logical value indicating if input is in ISO8601 format
}
\description{
This function is used to check if a --DTC variable is in ISO8601 format
}
\keyword{internal}
