% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_sec}
\alias{iso8601_sec}
\title{Format as ISO8601 seconds}
\usage{
iso8601_sec(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=iso8601_sec]{iso8601_sec()}} converts a character vector whose values represent seconds.
}
\keyword{internal}
