% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_truncate}
\alias{iso8601_truncate}
\title{Truncate a partial ISO8601 date-time}
\usage{
iso8601_truncate(x, empty_as_na = TRUE)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=iso8601_truncate]{iso8601_truncate()}} converts a character vector of ISO8601 dates, times or
date-times that might be partial and truncates the format by removing those
missing components.
}
\examples{
x <-
  c(
    "1999-01-01T15:20:01",
    "1999-01-01T15:20:-",
    "1999-01-01T15:-:-",
    "1999-01-01T-:-:-",
    "1999-01--T-:-:-",
    "1999----T-:-:-",
    "-----T-:-:-"
  )

sdtm.oak:::iso8601_truncate(x)

# With `empty_as_na = FALSE` empty strings are not replaced with `NA`
sdtm.oak:::iso8601_truncate("-----T-:-:-", empty_as_na = TRUE)
sdtm.oak:::iso8601_truncate("-----T-:-:-", empty_as_na = FALSE)

# Truncation only happens if missing components are the right most end,
# otherwise they remain unaltered.
sdtm.oak:::iso8601_truncate(
  c(
    "1999----T15:20:01",
    "1999-01-01T-:20:01",
    "1999-01-01T-:-:01",
    "1999-01-01T-:-:-"
  )
)

}
\keyword{internal}
