% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_two_digits}
\alias{iso8601_two_digits}
\title{Format as a ISO8601 two-digit number}
\usage{
iso8601_two_digits(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the same size as \code{x}.
}
\description{
\code{\link[=iso8601_two_digits]{iso8601_two_digits()}} converts a single digit or two digit number into a
two digit, 0-padded, number. Failing to parse the input as a two digit number
results in \code{NA}.
}
\keyword{internal}
