% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_year}
\alias{iso8601_year}
\title{Format as a ISO8601 four-digit year}
\usage{
iso8601_year(x, cutoff_2000 = 68L)
}
\arguments{
\item{x}{A character vector.}

\item{cutoff_2000}{A non-negative integer value. Two-digit years smaller or
equal to \code{cutoff_2000} are parsed as though starting with \code{20}, otherwise
parsed as though starting with \code{19}.}
}
\value{
A character vector.
}
\description{
\code{\link[=iso8601_year]{iso8601_year()}} converts a character vector whose values represent years to
four-digit years.
}
\keyword{internal}
