% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{months_abb_regex}
\alias{months_abb_regex}
\title{Regex for months' abbreviations}
\usage{
months_abb_regex(x = month.abb, case = c("any", "upper", "lower", "title"))
}
\arguments{
\item{x}{A character vector of three-letter month abbreviations. Default is
\code{month.abb}.}

\item{case}{A string scalar: \code{"any"}, if month abbreviations are to be
matched in any case; \code{"upper"}, to match uppercase abbreviations;
\code{"lower"}, to match lowercase; and, \code{"title"} to match title case.}
}
\value{
A regex as a string.
}
\description{
\code{\link[=months_abb_regex]{months_abb_regex()}} generates a regex that matches month abbreviations. For
finer control, the case can be specified with parameter \code{case}.
}
\keyword{internal}
