% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oak_id_vars.R
\name{oak_id_vars}
\alias{oak_id_vars}
\title{Raw dataset keys}
\usage{
oak_id_vars(extra_vars = NULL)
}
\arguments{
\item{extra_vars}{A character vector of extra column names to be appended to
the default names: oak_id, raw_source, and patient_number.}
}
\value{
A character vector of column names to be regarded
as keys in raw datasets.
}
\description{
\code{\link[=oak_id_vars]{oak_id_vars()}} is a helper function providing the variable (column) names to
be regarded as keys in \link[tibble:tibble-package]{tibbles} representing raw
datasets. By default, the set of names is
oak_id, raw_source, and patient_number. Extra variable names may be
indicated and passed in \code{extra_vars} which are appended to the default names.
}
