% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{parse_into_c_call}
\alias{parse_into_c_call}
\title{This function converts comma separated string into a character vector}
\usage{
parse_into_c_call(str_in)
}
\arguments{
\item{string}{A string with comma separated values.}
}
\value{
A character vector.
}
\description{
This function converts comma separated string into a character vector
}
\examples{
\dontrun{
str_in <- "a, b, c"
parse_into_c_call("a, b, c")

str_in <- NA_character_
parse_into_c_call(str_in)
}

}
\keyword{internal}
