% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{pseq}
\alias{pseq}
\title{Parallel sequence generation}
\usage{
pseq(from, to)
}
\arguments{
\item{from}{An integer vector. The starting value(s) of the sequence(s).}

\item{to}{An integer vector. The ending value(s) of the sequence(s).}
}
\value{
An integer vector.
}
\description{
\code{\link[=pseq]{pseq()}} is similar to \code{\link[=seq]{seq()}} but conveniently accepts integer vectors as
inputs to \code{from} and \code{to}, allowing for parallel generation of sequences.
The result is the union of the generated sequences.
}
\keyword{internal}
