% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{read_ct}
\alias{read_ct}
\title{Read in a controlled terminology}
\usage{
read_ct(file = stop("`file` must be specified"))
}
\arguments{
\item{file}{A path to a file containing a controlled terminology
specification data set. The following are expected of this file:
\itemize{
\item The file is expected to be a CSV file;
\item The file is expected to contain a first row of column names;
\item This minimal set of variables is expected: codelist_code, collected_value, term_synonyms, and term_value.
}}
}
\value{
A \link[tibble:tibble-package]{tibble} with a controlled terminology
specification.
}
\description{
\code{\link[=read_ct]{read_ct()}} imports a controlled terminology specification data set as a
\link[tibble:tibble-package]{tibble}.
}
