% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{read_spec}
\alias{read_spec}
\title{Read the specification file}
\usage{
read_spec(file)
}
\arguments{
\item{file}{The path to the specification file.}
}
\value{
A tibble with the specification.
}
\description{
Read the specification file
}
\examples{

\dontrun{
file <- "cm_sdtm_oak_spec_cdash.csv"
observed <- read_spec(file)
}

}
