% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_vars.R
\name{rec_vars}
\alias{rec_vars}
\title{Record-level key variables}
\usage{
rec_vars(domain)
}
\arguments{
\item{domain}{An SDTM domain in abbreviated form, e.g. \code{"DM"} (Demographics)
or \code{"CM"} (Concomitant Medications).}
}
\value{
A character vector of variable names.
}
\description{
\code{\link[=rec_vars]{rec_vars()}} returns the set of variable names that uniquely define
a record in a given SDTM domain.
}
\examples{
rec_vars("DM")

rec_vars("AE")

}
