% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode values}
\usage{
recode(x, from = unique(na.omit(x)), to = from, .no_match = x, .na = NA)
}
\arguments{
\item{x}{An atomic vector of values are to be recoded.}

\item{from}{A vector of values to be matched in \code{x} for recoding.}

\item{to}{A vector of values to be used as replacement for values in \code{from}.}

\item{.no_match}{Value to be used as replacement when cases in \code{from} are not
matched.}

\item{.na}{Value to be used to recode missing values.}
}
\value{
A vector of recoded values.
}
\description{
\code{\link[=recode]{recode()}} recodes values in \code{x} by matching elements in \code{from} onto values
in \code{to}.
}
\keyword{internal}
