% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{reg_matches}
\alias{reg_matches}
\title{\code{regmatches()} with \code{NA}}
\usage{
reg_matches(x, m, invert = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{m}{An object with match data.}

\item{invert}{A logical scalar. If \code{TRUE}, extract or replace the non-matched
substrings.}
}
\value{
A list of character vectors with the matched substrings, or \code{NA} if
matching failed.
}
\description{
\code{\link[=reg_matches]{reg_matches()}} is a thin wrapper around \code{\link[=regmatches]{regmatches()}} that returns
\code{NA} instead of \code{character(0)} when matching fails.
}
\keyword{internal}
