% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{regex_or}
\alias{regex_or}
\title{Utility function to assemble a regex of alternative patterns}
\usage{
regex_or(x, .open = FALSE, .close = FALSE)
}
\arguments{
\item{x}{A character vector of alternative patterns.}

\item{.open}{Whether the resulting regex should start with \code{"|"}.}

\item{.close}{Whether the resulting regex should end with \code{"|"}.}
}
\value{
A character scalar of the resulting regex.
}
\description{
\code{\link[=regex_or]{regex_or()}} takes a set of patterns and binds them with the Or (\code{"|"})
pattern for an easy regex of alternative patterns.
}
\examples{
# A regex for matching either "jan" or "feb"
sdtm.oak:::regex_or(c("jan", "feb"))

# Setting `.open` and/or `.close` to `TRUE` can be handy if this regex
# is to be combined into a larger regex.
paste0(sdtm.oak:::regex_or(c("jan", "feb"), .close = TRUE), r"{\d{2}}")

}
\keyword{internal}
