% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{remove_last_pipe}
\alias{remove_last_pipe}
\title{Remove the pipe operator from the last element of a character vector}
\usage{
remove_last_pipe(code_blocks)
}
\arguments{
\item{code_blocks}{A character vector.}
}
\value{
The character vector with the pipe operator removed from the last element.
}
\description{
Remove the pipe operator from the last element of a character vector
}
\keyword{internal}
