% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{rm_cnd_df}
\alias{rm_cnd_df}
\title{Remove the \code{cnd_df} class from a data frame}
\usage{
rm_cnd_df(dat)
}
\arguments{
\item{dat}{A data frame.}
}
\value{
The input \code{dat} without the \code{cnd_df} class.
}
\description{
This function removes the \code{cnd_df} class, along with its attributes, if
applicable.
}
