% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbj_vars.R
\name{sbj_vars}
\alias{sbj_vars}
\title{Subject-level key variables}
\usage{
sbj_vars()
}
\value{
A character vector of variable names.
}
\description{
\code{\link[=sbj_vars]{sbj_vars()}} returns the set of variable names that uniquely define
a subject.
}
\examples{
sbj_vars()

}
