% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.oak-package.R
\docType{package}
\name{sdtm.oak-package}
\alias{sdtm.oak}
\alias{sdtm.oak-package}
\title{sdtm.oak: SDTM Data Transformation Engine}
\description{
An EDC-agnostic and data standard-agnostic SDTM data transformation engine that provides a framework for modular programming of SDTM in R. Powered by metadata and standard specification sdtm.oak can potentially automate the conversion of raw clinical data to SDTM through standardized mapping algorithms. SDTM is one of the required standards for data submission to the FDA (U.S.) and PMDA (Japan). SDTM standards are implemented following the SDTM Implementation Guide as defined by CDISC \url{https://www.cdisc.org/standards/foundational/sdtmig}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/sdtm.oak/}
  \item Report bugs at \url{https://github.com/pharmaverse/sdtm.oak/issues/}
}

}
\author{
\strong{Maintainer}: Rammprasad Ganapathy \email{ganapathy.rammprasad@gene.com}

Authors:
\itemize{
  \item Adam Forys
  \item Edgar Manukyan
  \item Rosemary Li
  \item Preetesh Parikh
  \item Lisa Houterloot
  \item Yogesh Gupta
  \item Omar Garcia \email{ogcalderon@cdisc.org}
  \item Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})
  \item Kamil Sijko \email{kamil.sijko@ttsi.com.pl} (\href{https://orcid.org/0000-0002-2203-1065}{ORCID})
  \item Shiyu Chen \email{Shiyu.Chen@atorusresearch.com}
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Pfizer Inc [copyright holder, funder]
  \item Transition Technologies Science [copyright holder, funder]
}

}
\keyword{internal}
