% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{sdtm.oak}
\alias{sdtm.oak}
\title{An EDC and Data Standard agnostic SDTM data transformation engine that automates
the transformation of raw clinical data in ODM format to SDTM based on standard
mapping algorithms}
\description{
An EDC and Data Standard agnostic SDTM data transformation engine that automates
the transformation of raw clinical data in ODM format to SDTM based on standard
mapping algorithms
}
