% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R, R/sdtm_join.R
\name{sdtm_join}
\alias{sdtm_join}
\title{SDTM join}
\usage{
sdtm_join(raw_dat, tgt_dat = NULL, id_vars = oak_id_vars())

sdtm_join(raw_dat, tgt_dat = NULL, id_vars = oak_id_vars())
}
\arguments{
\item{raw_dat}{The raw dataset: a dataframe or a conditioned data frame. Must
include the variables passed in \code{id_vars}.}

\item{tgt_dat}{Target dataset: a data frame or a conditioned data frame to be
merged against \code{raw_dat} by the variables indicated in \code{id_vars}.}

\item{id_vars}{Key variables to be used in the join between the raw dataset
(\code{raw_dat}) and the target data set (\code{tgt_dat}).}
}
\value{
A data frame, or a conditioned data frame if at least one of the
input data sets is a conditioned data frame.

A data frame, or a conditioned data frame if, at least, one of the
input data sets is a conditioned data frame.
}
\description{
\code{\link[=sdtm_join]{sdtm_join()}} is a special join between a raw data set and a target data
set. This function supports conditioned data frames.

\code{\link[=sdtm_join]{sdtm_join()}} is a special join between a raw data set and a target data
set. This function supports conditioned data frames.
}
\keyword{internal}
