% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_var_call.R
\name{sdtm_str_split}
\alias{sdtm_str_split}
\title{Split a dataset string}
\usage{
sdtm_str_split(domain_dataset, max_length_out = 200L)
}
\arguments{
\item{domain_dataset}{The dataset to be split}

\item{max_length_out}{The maximum length of output string (default is 200)}
}
\value{
A dataset with the split strings
}
\description{
Split a dataset string
}
