% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_var_call.R
\name{single_str_spilt}
\alias{single_str_spilt}
\title{Split a single string}
\usage{
single_str_spilt(result, item, sep, max_length_out)
}
\arguments{
\item{result}{A list to store the result}

\item{item}{The item to be added to the result}

\item{sep}{The separator used to split the string}

\item{max_length_out}{The maximum length of the output string}
}
\value{
A list with the split strings
}
\description{
Split a single string
}
