% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_var_call.R
\name{split_var}
\alias{split_var}
\title{Split a variable}
\usage{
split_var(string, max_length_out = 200L)
}
\arguments{
\item{string}{The string to be split}

\item{max_length_out}{The maximum length of the output string (default is 200)}
}
\value{
A list with the split strings
}
\description{
Split a variable
}
