% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dttm_fmt.R
\name{str_to_anycase}
\alias{str_to_anycase}
\title{Generate case insensitive regexps}
\usage{
str_to_anycase(x)
}
\arguments{
\item{x}{A character vector of strings consisting of word characters.}
}
\value{
A character vector.
}
\description{
\code{\link[=str_to_anycase]{str_to_anycase()}} takes a character vector of word strings as input, and
generates regular expressions that express that match in any case.
}
\keyword{internal}
