% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{style_the_code}
\alias{style_the_code}
\title{Style the code}
\usage{
style_the_code(code_by_topics, domain)
}
\arguments{
\item{code_by_topics}{A list of character vectors.}

\item{domain}{The SDTM domain to generate the code for.}
}
\value{
The styled code as a string.
}
\description{
This function styles the code using the styler package and adds the necessary
templates to the code (e.g. cm_template_prefix, cm_template_suffix).
}
\keyword{internal}
