% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{tbl_sum.cnd_df}
\alias{tbl_sum.cnd_df}
\title{Print}
\usage{
\method{tbl_sum}{cnd_df}(x, ...)
}
\arguments{
\item{x}{A conditioned tibble of class \code{cnd_df}.}

\item{...}{Additional arguments passed to the default print method.}
}
\description{
Blah
}
