% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{zero_pad_whole_number}
\alias{zero_pad_whole_number}
\title{Convert an integer to a zero-padded character vector}
\usage{
zero_pad_whole_number(x, n = 2L)
}
\arguments{
\item{x}{An integer vector.}

\item{n}{Number of digits in the output, including zero padding.}
}
\value{
A character vector.
}
\description{
\code{\link[=zero_pad_whole_number]{zero_pad_whole_number()}} takes non-negative integer values and converts
them to character with zero padding. Negative numbers and numbers greater
than the width specified by the number of digits \code{n} are converted to \code{NA}.
}
\keyword{internal}
