% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_utils.R
\name{assert_dtc_fmt}
\alias{assert_dtc_fmt}
\title{Assert date time character formats}
\usage{
assert_dtc_fmt(fmt)
}
\arguments{
\item{fmt}{A character vector.}
}
\description{
\code{\link[=assert_dtc_fmt]{assert_dtc_fmt()}} takes a character vector of date/time formats and checks if
the formats are supported, meaning it checks if they are one of the formats
listed in column \code{fmt} of \link{dtc_formats}, failing with an error otherwise.
}
\keyword{internal}
