% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\alias{assign_no_ct}
\alias{assign_ct}
\title{Derive an SDTM variable}
\usage{
assign_no_ct(
  raw_dataset,
  raw_variable,
  target_sdtm_variable,
  target_dataset = raw_dataset,
  merge_to_topic_by = NULL
)

assign_ct(
  raw_dataset,
  raw_variable,
  target_sdtm_variable,
  target_dataset = raw_dataset,
  merge_to_topic_by = NULL,
  study_ct = NULL,
  target_sdtm_variable_codelist_code = NULL
)
}
\arguments{
\item{raw_dataset}{The raw dataset.}

\item{raw_variable}{The raw variable.}

\item{target_sdtm_variable}{The target SDTM variable.}

\item{target_dataset}{Target dataset. By default the same as \code{raw_dataset}.}

\item{merge_to_topic_by}{If \code{target_dataset} is different than \code{raw_dataset},
then this parameter defines keys to use in the join between \code{raw_dataset}
and \code{target_dataset}.}

\item{study_ct}{Study controlled terminology specification.}

\item{target_sdtm_variable_codelist_code}{A codelist code indicating which
subset of the controlled terminology to apply in the derivation.}
}
\value{
The target dataset with the derived variable \code{target_sdtm_variable}.

The target dataset with the derived variable \code{target_sdtm_variable}.
}
\description{
\itemize{
\item \code{\link[=assign_no_ct]{assign_no_ct()}} maps a variable in a source dataset to a target SDTM
variable that has no terminology restrictions.
\item \code{\link[=assign_ct]{assign_ct()}} maps a variable in a source dataset to a target SDTM variable
following controlled terminology recoding.
}

\code{\link[=assign_ct]{assign_ct()}} maps a variable in a source dataset to a target SDTM variable
following controlled terminology recoding.
}
\examples{
study_ct <-
  tibble::tibble(
    codelist_code = rep("C66729", 8L),
    term_code = c(
      "C28161",
      "C38210",
      "C38222",
      "C38223",
      "C38287",
      "C38288",
      "C38305",
      "C38311"
    ),
    CodedData = c(
      "INTRAMUSCULAR",
      "EPIDURAL",
      "INTRA-ARTERIAL",
      "INTRA-ARTICULAR",
      "OPHTHALMIC",
      "ORAL",
      "TRANSDERMAL",
      "UNKNOWN"
    ),
    term_value = CodedData,
    collected_value = c(
      "IM (Intramuscular)",
      "EP (Epidural)",
      "IA (Intra-arterial)",
      "IJ (Intra-articular)",
      "OP (Ophthalmic)",
      "PO (Oral)",
      "DE (Transdermal)",
      "Unknown"
    ),
    term_preferred_term = c(
      "Intramuscular Route of Administration",
      "Epidural Route of Administration",
      "Intraarterial Route of Administration",
      "Intraarticular Route of Administration",
      "Ophthalmic Route of Administration",
      "Oral Route of Administration",
      "Transdermal Route of Administration",
      "Unknown Route of Administration"
    ),
    term_synonyms = c(rep(NA, 5L), "Intraoral Route of Administration; PO", NA, NA),
    raw_codelist = rep("ROUTE_CV1", 8L)
  )

md1 <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    PATIENT_NUM = 101:114,
    MDRTE = c(
      "PO (Oral)",
      "PO (Oral)",
      NA_character_,
      "PO",
      "Intraoral Route of Administration",
      "PO (Oral)",
      "IM (Intramuscular)",
      "IA (Intra-arterial)",
      "",
      "Non-standard",
      "random_value",
      "IJ (Intra-articular)",
      "TRANSDERMAL",
      "OPHTHALMIC"
    )
  )

assign_ct(
  raw_dataset = md1,
  raw_variable = "MDRTE",
  study_ct = study_ct,
  target_sdtm_variable = "CMROUTE",
  target_sdtm_variable_codelist_code = "C66729"
  )

cm_inter <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    PATIENT_NUM = 101:114,
    CMTRT = c(
      "BABY ASPIRIN",
      "CORTISPORIN",
      "ASPIRIN",
      "DIPHENHYDRAMINE HCL",
      "PARCETEMOL",
      "VOMIKIND",
      "ZENFLOX OZ",
      "AMITRYPTYLINE",
      "BENADRYL",
      "DIPHENHYDRAMINE HYDROCHLORIDE",
      "TETRACYCLINE",
      "BENADRYL",
      "SOMINEX",
      "ZQUILL"
    ),
    CMINDC = c(
      NA,
      "NAUSEA",
      "ANEMIA",
      "NAUSEA",
      "PYREXIA",
      "VOMITINGS",
      "DIARHHEA",
      "COLD",
      "FEVER",
      "LEG PAIN",
      "FEVER",
      "COLD",
      "COLD",
      "PAIN"
    ),
    CMROUTE = c(
      "ORAL",
      "ORAL",
      NA,
      "ORAL",
      "ORAL",
      "ORAL",
      "INTRAMUSCULAR",
      "INTRA-ARTERIAL",
      NA,
      "NON-STANDARD",
      "RANDOM_VALUE",
      "INTRA-ARTICULAR",
      "TRANSDERMAL",
      "OPHTHALMIC"
    )
  )

assign_ct(
  raw_dataset = md1,
  raw_variable = "MDRTE",
  study_ct = study_ct,
  target_sdtm_variable = "CMROUTE",
  target_sdtm_variable_codelist_code = "C66729",
  target_dataset = cm_inter,
  merge_to_topic_by = c("oak_id","raw_source","PATIENT_NUM")
  )

study_ct <-
  tibble::tibble(
    codelist_code = rep("C66729", 8L),
    term_code = c(
      "C28161",
      "C38210",
      "C38222",
      "C38223",
      "C38287",
      "C38288",
      "C38305",
      "C38311"
    ),
    CodedData = c(
      "INTRAMUSCULAR",
      "EPIDURAL",
      "INTRA-ARTERIAL",
      "INTRA-ARTICULAR",
      "OPHTHALMIC",
      "ORAL",
      "TRANSDERMAL",
      "UNKNOWN"
    ),
    term_value = CodedData,
    collected_value = c(
      "IM (Intramuscular)",
      "EP (Epidural)",
      "IA (Intra-arterial)",
      "IJ (Intra-articular)",
      "OP (Ophthalmic)",
      "PO (Oral)",
      "DE (Transdermal)",
      "Unknown"
    ),
    term_preferred_term = c(
      "Intramuscular Route of Administration",
      "Epidural Route of Administration",
      "Intraarterial Route of Administration",
      "Intraarticular Route of Administration",
      "Ophthalmic Route of Administration",
      "Oral Route of Administration",
      "Transdermal Route of Administration",
      "Unknown Route of Administration"
    ),
    term_synonyms = c(rep(NA, 5L), "Intraoral Route of Administration; PO", NA, NA),
    raw_codelist = rep("ROUTE_CV1", 8L)
  )

md1 <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    PATIENT_NUM = 101:114,
    MDRTE = c(
      "PO (Oral)",
      "PO (Oral)",
      NA_character_,
      "PO",
      "Intraoral Route of Administration",
      "PO (Oral)",
      "IM (Intramuscular)",
      "IA (Intra-arterial)",
      "",
      "Non-standard",
      "random_value",
      "IJ (Intra-articular)",
      "TRANSDERMAL",
      "OPHTHALMIC"
    )
  )

assign_ct(
  raw_dataset = md1,
  raw_variable = "MDRTE",
  study_ct = study_ct,
  target_sdtm_variable = "CMROUTE",
  target_sdtm_variable_codelist_code = "C66729"
  )

cm_inter <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    PATIENT_NUM = 101:114,
    CMTRT = c(
      "BABY ASPIRIN",
      "CORTISPORIN",
      "ASPIRIN",
      "DIPHENHYDRAMINE HCL",
      "PARCETEMOL",
      "VOMIKIND",
      "ZENFLOX OZ",
      "AMITRYPTYLINE",
      "BENADRYL",
      "DIPHENHYDRAMINE HYDROCHLORIDE",
      "TETRACYCLINE",
      "BENADRYL",
      "SOMINEX",
      "ZQUILL"
    ),
    CMINDC = c(
      NA,
      "NAUSEA",
      "ANEMIA",
      "NAUSEA",
      "PYREXIA",
      "VOMITINGS",
      "DIARHHEA",
      "COLD",
      "FEVER",
      "LEG PAIN",
      "FEVER",
      "COLD",
      "COLD",
      "PAIN"
    ),
    CMROUTE = c(
      "ORAL",
      "ORAL",
      NA,
      "ORAL",
      "ORAL",
      "ORAL",
      "INTRAMUSCULAR",
      "INTRA-ARTERIAL",
      NA,
      "NON-STANDARD",
      "RANDOM_VALUE",
      "INTRA-ARTICULAR",
      "TRANSDERMAL",
      "OPHTHALMIC"
    )
  )

assign_ct(
  raw_dataset = md1,
  raw_variable = "MDRTE",
  study_ct = study_ct,
  target_sdtm_variable = "CMROUTE",
  target_sdtm_variable_codelist_code = "C66729",
  target_dataset = cm_inter,
  merge_to_topic_by = c("oak_id","raw_source","PATIENT_NUM")
  )


}
