% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\alias{assign_no_ct}
\alias{assign_ct}
\title{Derive an SDTM variable}
\usage{
assign_no_ct(
  raw_dat,
  raw_var,
  tgt_var,
  tgt_dat = NULL,
  id_vars = oak_id_vars()
)

assign_ct(
  raw_dat,
  raw_var,
  tgt_var,
  ct,
  cl,
  tgt_dat = NULL,
  id_vars = oak_id_vars()
)
}
\arguments{
\item{raw_dat}{The raw dataset.}

\item{raw_var}{The raw variable.}

\item{tgt_var}{The target SDTM variable.}

\item{tgt_dat}{Target dataset.}

\item{id_vars}{If \code{target_dataset} is different than \code{raw_dataset},
then this parameter defines keys to use in the join between \code{raw_dataset}
and \code{target_dataset}.}

\item{ct}{Study controlled terminology specification.}

\item{cl}{A codelist code indicating which
subset of the controlled terminology to apply in the derivation.}
}
\value{
The target dataset with the derived variable \code{target_sdtm_variable}.
}
\description{
\itemize{
\item \code{\link[=assign_no_ct]{assign_no_ct()}} maps a variable in a source dataset to a target SDTM
variable that has no terminology restrictions.
\item \code{\link[=assign_ct]{assign_ct()}} maps a variable in a source dataset to a target SDTM variable
following controlled terminology recoding.
}
}
\examples{

md1 <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    patient_number = 101:114,
    MDIND = c( "NAUSEA", "NAUSEA", "ANEMIA", "NAUSEA", "PYREXIA",
    "VOMITINGS", "DIARHHEA", "COLD",
    "FEVER", "LEG PAIN", "FEVER", "COLD", "COLD", "PAIN"
    )
  )

assign_no_ct(
  raw_dat = md1,
  raw_var = "MDIND",
  tgt_var = "CMINDC",
  )

cm_inter <-
  tibble::tibble(
    oak_id = 1:14,
    raw_source = "MD1",
    patient_number = 101:114,
    CMTRT = c(
      "BABY ASPIRIN",
      "CORTISPORIN",
      "ASPIRIN",
      "DIPHENHYDRAMINE HCL",
      "PARCETEMOL",
      "VOMIKIND",
      "ZENFLOX OZ",
      "AMITRYPTYLINE",
      "BENADRYL",
      "DIPHENHYDRAMINE HYDROCHLORIDE",
      "TETRACYCLINE",
      "BENADRYL",
      "SOMINEX",
      "ZQUILL"
    ),
    CMROUTE = c(
      "ORAL",
      "ORAL",
      NA,
      "ORAL",
      "ORAL",
      "ORAL",
      "INTRAMUSCULAR",
      "INTRA-ARTERIAL",
      NA,
      "NON-STANDARD",
      "RANDOM_VALUE",
      "INTRA-ARTICULAR",
      "TRANSDERMAL",
      "OPHTHALMIC"
    )
  )

assign_ct(
  raw_dat = md1,
  raw_var = "MDIND",
  tgt_var = "CMINDC",
  tgt_dat = cm_inter,
  id_vars = c("oak_id","raw_source","patient_number")
  )

}
