% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_utils.R
\name{complete_capture_matrix}
\alias{complete_capture_matrix}
\title{Complete a capture matrix}
\usage{
complete_capture_matrix(m)
}
\arguments{
\item{m}{A character matrix that might be missing one or more of the
following columns: \code{year}, \code{mon}, \code{mday}, \code{hour}, \code{min} or \code{sec}.}
}
\value{
A character matrix that contains the columns \code{year}, \code{mon}, \code{mday},
\code{hour}, \code{min} and \code{sec}. Any other existing columns are dropped.
}
\description{
\code{\link[=complete_capture_matrix]{complete_capture_matrix()}} completes the missing, if any, columns of the
capture matrix.
}
\keyword{internal}
