% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_seq.R
\name{derive_seq}
\alias{derive_seq}
\title{Derive the sequence number (\code{--SEQ}) variable}
\usage{
derive_seq(tgt_dat, tgt_var, id_vars, start_at = 1L)
}
\arguments{
\item{tgt_dat}{The target dataset, a data frame.}

\item{tgt_var}{The target SDTM variable: a single string indicating the name
of the sequence number (\code{--SEQ}) variable, e.g. \code{"DSSEQ"}. Note that
supplying a name not ending in \code{"SEQ"} will raise a warning.}

\item{id_vars}{Either a character vector of identifier variables.
Default is the set of variables returned by \code{\link[=oak_id_vars]{oak_id_vars()}}.}

\item{start_at}{The sequence numbering starts at this value (default is \code{1}).}
}
\value{
Returns the data frame supplied in \code{tgt_dat} with the newly derived
variable, i.e. the sequence number (\code{--SEQ}), whose name is that passed in
\code{tgt_var}. This variable is of type integer.
}
\description{
\code{\link[=derive_seq]{derive_seq()}} creates a new identifier variable: the sequence number
(\code{--SEQ}).

This function adds a newly derived variable to \code{tgt_dat}, namely the sequence
number (\code{--SEQ}) whose name is the one provided in \code{tgt_var}. An integer
sequence is generated that uniquely identifies each record within the domain.

Prior to the derivation of \code{tgt_var}, the data frame \code{tgt_dat} is sorted
according to grouping variables indicated in \code{id_vars}.
}
