% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_seq.R
\name{derive_seq}
\alias{derive_seq}
\title{Derive the sequence number (\code{--SEQ}) variable}
\usage{
derive_seq(
  tgt_dat,
  tgt_var,
  id_vars = oak_id_vars(),
  sort_vars = purrr::set_names(rlang::rep_along(id_vars, "asc"), id_vars),
  start_at = 1L
)
}
\arguments{
\item{tgt_dat}{The target dataset, a data frame.}

\item{tgt_var}{The target SDTM variable: a single string indicating the name
of the sequence number (\code{--SEQ}) variable, e.g. \code{"DSSEQ"}. Note that supplying
a name not ending in \code{"SEQ"} will raise a warning.}

\item{id_vars}{Either a character vector of identifier variables or \code{NULL}.
Default is the set of variables returned by \code{\link[=oak_id_vars]{oak_id_vars()}}.}
}
\value{
Returns the data frame supplied in \code{tgt_dat} with the newly derived
variable, i.e. the sequence number (\code{--SEQ}) whose name is that passed in
\code{tgt_var}. This variable is of type integer.
}
\description{
\code{\link[=derive_seq]{derive_seq()}} creates a new identifier variable: the sequence number
(\code{--SEQ}).

This function adds a newly derived variable to \code{tgt_dat}, namely the sequence
number (\code{--SEQ}) whose name is \code{tgt_var}. An integer sequence is generated
that, when used in combination with the identifier of the subject, uniquely
identifies each record within the domain.

If \code{tgt_dat} pertains a domain that does not contain subject data, then
passing \code{NULL} to \code{vars_id} will ensure that every row in \code{tgt_dat} gets a
unique integer number in the column named \code{tgt_var}.
}
\examples{
#


}
