% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sdtm_supp.R
\name{gen_sdtm_supp}
\alias{gen_sdtm_supp}
\title{Function to generate final SDTM domain and supplemental domain SUPP--}
\usage{
gen_sdtm_supp(sdtm_dataset, idvar = NULL, spec, qnam_var, label_var, orig_var)
}
\arguments{
\item{sdtm_dataset}{SDTM output used to split supplemental domains.}

\item{idvar}{Variable name for IDVAR variable.}

\item{spec}{User-defined data frame of specifications which should contain
\code{qnam_var}, \code{label_var} and \code{orig_var}.}

\item{qnam_var}{Variable name in user-defined \code{spec} for QNAM variable.}

\item{label_var}{Variable name in user-defined \code{spec} for QLABEL variable.}

\item{orig_var}{Variable name in user-defined \code{spec} for QORIG variable.}
}
\value{
List of SDTM domain with suppquals dropped and corresponding supplemental domain.
}
\description{
Function to generate final SDTM domain and supplemental domain SUPP--
}
\examples{
path <- here::here("inst")
dm <- readRDS(paste0(path, "/domain/dm.rds"))
spec <- read.csv(paste0(path, "/spec/suppqual_spec.csv"))
final <- gen_sdtm_supp(dm, idvar = NULL, spec = spec, qnam_var = "Variable", label_var = "Label", orig_var = "Origin")
}
