% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_code.R
\name{generate_one_var_code}
\alias{generate_one_var_code}
\title{Generate the code for one variable}
\usage{
generate_one_var_code(spec_var, last_var = FALSE)
}
\arguments{
\item{spec_var}{The specification for one variable.}

\item{last_var}{Logical indicating if this is the last variable in the domain.}
}
\value{
The code for the variable as a string.
}
\description{
Generate the code for one variable
}
\keyword{internal}
