% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardcode.R
\name{harcode}
\alias{harcode}
\alias{hardcode_no_ct}
\alias{hardcode_ct}
\title{Derive an SDTM variable with a hardcoded value}
\usage{
hardcode_no_ct(
  raw_dataset,
  raw_variable,
  target_sdtm_variable,
  target_hardcoded_value,
  target_dataset = NULL,
  id_vars = NULL
)

hardcode_ct(
  raw_dataset,
  raw_variable,
  target_sdtm_variable,
  target_hardcoded_value,
  target_dataset = NULL,
  id_vars = NULL,
  study_ct,
  target_sdtm_variable_codelist_code
)
}
\arguments{
\item{raw_dataset}{The raw dataset.}

\item{raw_variable}{The raw variable.}

\item{target_sdtm_variable}{The target SDTM variable.}

\item{target_hardcoded_value}{Hardcoded value.}

\item{target_dataset}{Target dataset. By default the same as \code{raw_dataset}.}

\item{id_vars}{If \code{target_dataset} is different than \code{raw_dataset},
then this parameter defines keys to use in the join between \code{raw_dataset}
and \code{target_dataset}.}

\item{study_ct}{Study controlled terminology specification.}

\item{target_sdtm_variable_codelist_code}{A codelist code indicating which
subset of the controlled terminology to apply in the derivation.}
}
\value{
The target dataset with the derived variable \code{target_sdtm_variable}.
}
\description{
\itemize{
\item \code{\link[=hardcode_no_ct]{hardcode_no_ct()}} maps a hardcoded value to a target SDTM variable that has
no terminology restrictions.
\item \code{\link[=hardcode_ct]{hardcode_ct()}} maps a hardcoded value to a target SDTM variable with
controlled terminology recoding.
}
}
\examples{
MD1 <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~MDRAW,
    1L, "MD1", 101L, "BABY ASPIRIN",
    2L, "MD1", 102L, "CORTISPORIN",
    3L, "MD1", 103L, NA_character_,
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS".
hardcode_no_ct(
  raw_dataset = MD1,
  raw_variable = "MDRAW",
  target_sdtm_variable = "CMCAT",
  target_hardcoded_value = "GENERAL CONCOMITANT MEDICATIONS"
)

CM_INTER <-
  tibble::tribble(
    ~oak_id, ~raw_source, ~patient_number, ~CMTRT, ~CMINDC,
    1L, "MD1", 101L, "BABY ASPIRIN", NA,
    2L, "MD1", 102L, "CORTISPORIN", "NAUSEA",
    3L, "MD1", 103L, "ASPIRIN", "ANEMIA",
    4L, "MD1", 104L, "DIPHENHYDRAMINE HCL", "NAUSEA",
    5L, "MD1", 105L, "PARACETAMOL", "PYREXIA"
  )

# Derive a new variable `CMCAT` by overwriting `MDRAW` with the
# hardcoded value "GENERAL CONCOMITANT MEDICATIONS" with a prior join to
# `target_dataset`.
hardcode_no_ct(
  raw_dataset = MD1,
  raw_variable = "MDRAW",
  target_sdtm_variable = "CMCAT",
  target_hardcoded_value = "GENERAL CONCOMITANT MEDICATIONS",
  target_dataset = CM_INTER,
  id_vars = c("oak_id", "raw_source", "patient_number")
)

}
