% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_seq.R
\name{is_seq_name}
\alias{is_seq_name}
\title{Is it a --SEQ variable name}
\usage{
is_seq_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A logical vector.
}
\description{
\code{\link[=is_seq_name]{is_seq_name()}} returns which variable names end in \code{"SEQ"}.
}
\keyword{internal}
