% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_mon}
\alias{iso8601_mon}
\title{Format as a ISO8601 month}
\usage{
iso8601_mon(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=iso8601_mon]{iso8601_mon()}} converts a character vector whose values represent numeric
or abbreviated month names to zero-padded numeric months.
}
\keyword{internal}
