% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{iso8601_truncate}
\alias{iso8601_truncate}
\title{Truncate a partial ISO8601 date-time}
\usage{
iso8601_truncate(x, empty_as_na = TRUE)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
\code{\link[=iso8601_truncate]{iso8601_truncate()}} converts a character vector of ISO8601 dates, times or
date-times that might be partial and truncates the format by removing those
missing components.
}
\keyword{internal}
