% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd_df.R
\name{rm_cnd_df}
\alias{rm_cnd_df}
\title{Remove the \code{cnd_df} class from a data frame}
\usage{
rm_cnd_df(dat)
}
\arguments{
\item{dat}{A data frame.}
}
\value{
The input \code{dat} without the \code{cnd_df} class and associated attributes.
}
\description{
This function removes the \code{cnd_df} class, along with its attributes, if
applicable.
}
\seealso{
\code{\link[=new_cnd_df]{new_cnd_df()}}, \code{\link[=is_cnd_df]{is_cnd_df()}}, \code{\link[=get_cnd_df_cnd]{get_cnd_df_cnd()}},
\code{\link[=get_cnd_df_cnd_sum]{get_cnd_df_cnd_sum()}}.
}
\keyword{internal}
