% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.oak-package.R
\docType{package}
\name{sdtm.oak-package}
\alias{sdtm.oak}
\alias{sdtm.oak-package}
\title{sdtm.oak: SDTM Data Transformation Engine}
\description{
An EDC and Data Standard-agnostic SDTM data transformation engine designed for SDTM programming in R. Powered by metadata sdtm.oak can automate the conversion of raw clinical data to SDTM through standardized mapping algorithms. SDTM is one of the required standards for data submission to FDA (U.S.) and PMDA (Japan). SDTM standards are implemented in accordance with the SDTM Implementation guide as defined by CDISC \url{https://www.cdisc.org/standards/foundational/sdtmig}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/sdtm.oak/}
  \item \url{https://github.com/pharmaverse/sdtm.oak}
  \item Report bugs at \url{https://github.com/pharmaverse/sdtm.oak/issues}
}

}
\author{
\strong{Maintainer}: Rammprasad Ganapathy \email{ganapathy.rammprasad@gene.com}

Authors:
\itemize{
  \item Adam Forys
  \item Edgar Manukyan
  \item Rosemary Li
  \item Preetesh Parikh
  \item Lisa Houterloot
  \item Yogesh Gupta
  \item Omar Garcia \email{ogcalderon@cdisc.org}
  \item Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Pfizer Inc [copyright holder, funder]
}

}
\keyword{internal}
