% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.oak-package.R
\docType{package}
\name{sdtm.oak-package}
\alias{sdtm.oak}
\alias{sdtm.oak-package}
\title{sdtm.oak: SDTM Data Transformation Engine}
\description{
An EDC and Data Standard agnostic SDTM data transformation engine that automates the transformation of raw clinical data in ODM format to SDTM based on standard mapping algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/sdtm.oak/}
  \item \url{https://github.com/pharmaverse/sdtm.oak}
  \item Report bugs at \url{https://github.com/pharmaverse/sdtm.oak/issues}
}

}
\author{
\strong{Maintainer}: Omar Garcia \email{ogcalderon@cdisc.org}

Authors:
\itemize{
  \item Rammprasad Ganapathy
  \item Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Pfizer Inc [copyright holder, funder]
}

}
\keyword{internal}
