% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtc_create_iso8601.R
\name{yy_to_yyyy}
\alias{yy_to_yyyy}
\title{Convert two-digit to four-digit years}
\usage{
yy_to_yyyy(x, cutoff_2000 = 68L)
}
\arguments{
\item{x}{An integer vector of years.}

\item{cutoff_2000}{An integer value. Two-digit years smaller or equal to
\code{cutoff_2000} are parsed as though starting with \code{20}, otherwise parsed as
though starting with \code{19}.}
}
\value{
An integer vector.
}
\description{
\code{\link[=yy_to_yyyy]{yy_to_yyyy()}} converts two-digit years to four-digit years.
}
\keyword{internal}
